Распределить классы по package
