+++ src/http/handler/PrioritizedHandler.java
@@ -1,4 +1,4 @@
-package http;
+package http.handler;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
@@ -6,17 +6,17 @@ import manager.TaskManager;
 
 import java.io.IOException;
 
-public class HistoryHandler extends BaseHttpHandler implements HttpHandler {
+public class PrioritizedHandler extends BaseHttpHandler implements HttpHandler {
 
-    public HistoryHandler(TaskManager taskManager) {
+    public PrioritizedHandler(TaskManager taskManager) {
         super(taskManager);
     }
 
     @Override
     public void handle(HttpExchange exchange) throws IOException {
-        if (exchange.getRequestURI().getPath().equals("/history")) {
+        if (exchange.getRequestURI().getPath().equals("/prioritized")) {
             if (exchange.getRequestMethod().equals("GET")) {
-                String response = gson.toJson(manager.getHistory());
+                String response = gson.toJson(manager.getPrioritizedTasks());
                 sendText(exchange, response, 200);
             }
         }
